[OP-28] fixed travis warnings and tests instrument_drivers
