+++ pytest.ini
@@ -11,3 +11,4 @@ norecursedirs = .svn _build tmp* .git docs untitled* deprecated* build dist .cac
 # ignore deprecation warnings from external packages
 filterwarnings =
     ignore:.*invalid escape sequence.*:DeprecationWarning
+    ignore:.*No module named '_tifffile'*:UserWarning
